Switch to new tar and Makefile
Unable to create two types of files on macOS, so removed them from the
test-directory.tar that was previously part of this commit path.

They are "nod-c" and "nod-b" using two types of files I've never seen
before.

The makefile is tested on macOS as sudo and non-root user. Helpful step
echo'd at the end tells you what to run next.

Makefile also will untar and not fail if there are any errors by
prefixing the step with "-"